+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-04-18 09:32:21
-// Commit: test(theme): implement widget integration tests for theme
+// Updated at: 2026-04-18 12:06:29
+// Commit: chore(theme): update dependencies and cleanup unused imports in theme
 
 class themeComponentState {
-  final String commitRef = '505';
-  final String lastMessage = 'test(theme): implement widget integration tests for theme';
-  final DateTime timestamp = DateTime.parse('2026-04-18T09:32:21+05:00');
+  final String commitRef = '511';
+  final String lastMessage = 'chore(theme): update dependencies and cleanup unused imports in theme';
+  final DateTime timestamp = DateTime.parse('2026-04-18T12:06:29+05:00');
 }